And Couple Name Changes
And Couple Name Changes

+++ FashionNewsFeed/FashionNewsFeed/FCCategory.h
@@ -1,13 +1,21 @@
+//
+//  FCCategory.h
+//  FashionNewsFeed
+//
+//  Created by Anton Chugunov on 27.03.15.
+//  Copyright (c) 2015 Anton Chugunov. All rights reserved.
+//
+
 #import <Foundation/Foundation.h>
 
 @interface FCCategory : NSObject {
-    int categoryId;
+    NSUInteger categoryId;
     NSString* categoryName;
 }
 
-@property (nonatomic) int categoryId;
+@property (nonatomic) NSUInteger categoryId;
 @property (strong, nonatomic) NSString* categoryName;
 
-- (id)initWithNameAndId:(NSString*)categoryName categoryId:(int)categoryId;
+- (id)initWithNameAndId:(NSString*)categoryName categoryId:(NSUInteger)categoryId;
 
 @end
